requirements.txt
kafka-python
textblob
newsapi-python